backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==11.0.0
gunicorn==23.0.0
whitenoise==6.5.0
django-storages[boto3]==1.14.1
boto3==1.28.0
